+++ modules/admin/src/messages/de/admin.php
@@ -149,4 +149,6 @@ return [
     'login_securetoken_mail' => '<h1>LUYA Sicherheitscode</h1><p>Verwenden Sie den folgenden Sicherheitscode für den Zugriff auf die Administration der Website {url}:</p><p><strong>{token}</strong></p>',
     'login_async_token_error' => 'Der eingegeben Sicherheitscode ist falsch',
     'login_async_token_globalerror' => 'Ein Globaler-Fehler ist enstanden. Bitte kontaktieren Sie Ihren Seitenbetreiber.',
+    'ngrest_crud_csv_export_btn' => 'CSV generieren',
+    'ngrest_crud_csv_export_btn_dl' => 'CSV herunterladen',
 ];