+++ modules/admin/src/messages/en/admin.php
@@ -149,4 +149,6 @@ return [
     'login_securetoken_mail' => '<h1>LUYA Security code</h1><p>Use the following security code to login into your administration interface on {url}:</p><p><strong>{token}</strong></p>',
     'login_async_token_error' => 'The provided security code is wrong.',
     'login_async_token_globalerror' => 'A Server Error occured, please contact your Site administrator.',
+    'ngrest_crud_csv_export_btn' => 'Generate CSV',
+    'ngrest_crud_csv_export_btn_dl' => 'Download CSV',
 ];